Added styling to Dashboard divs, added header and footer to AddRecipes but could not implement styling from main css, will need help
